pyproject.toml
[tool.poetry]
name = "video_compressing"
version = "0.1.4"
description = "Package to compress multiple large .MOV or .MP4 videos into a single light .mp4 video."
authors = ["gabriel-melki"]
readme = "README.md"
packages = [
    { include = "video_compressing" }
]
homepage = "https://github.com/Gabriel-melki/video_compressing"
repository = "https://github.com/Gabriel-melki/video_compressing"
documentation = "https://github.com/Gabriel-melki/video_compressing#readme"

[tool.poetry.dependencies]
python = "^3.12.0"
requests = "^2.32.3"
subprocess32 = "^3.5.4"
tqdm = "^4.63.0"
numpy = "^2.1.3"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "7.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"